==> scripts/06_annotations_with_Harmony_integration.Rmd <==
---
title: "annotations before doublet removal to remove noise clusters"
author: "Kimberly Olney, PhD"
date: "04/19/2024"
output:
  html_document:
    theme: cerulean
    toc: true
    toc_float: true
editor_options: 
  chunk_output_type: inline
---


```{r setup, include=FALSE}
knitr::opts_chunk$set(root.dir = ".", echo=TRUE, warning=FALSE, message=FALSE)
```

# Libraris, paths, colors
```{r echo=FALSE, message=FALSE}
source(here::here("/research/labs/neurology/fryer/m239830/PMI/scripts", "file_paths_and_colours.R"))
treatment <- "PMI_integrated"
color.panel <- dittoColors()
metadata <- subset(metadata, sampleID != "P_60" & sampleID != "B_6")
metadata$sampleID <- factor(metadata$sampleID, levels = c(metadata$sampleID))
```

# Read in object
```{r read_object}
# read object
dataObject <- readRDS(file = paste0("../rObjects/PMI_unannotated_integrated.rds"))
# inspect
dataObject
DefaultAssay(dataObject) <- "RNA"
dataObject <- NormalizeData(dataObject)
dataObject <- ScaleData(dataObject)
```

```{r}
DimPlot(dataObject,
        group.by = "strain")

DimPlot(dataObject,
        group.by = "sex")

DimPlot(dataObject,
        group.by = "time")

DimPlot(dataObject,
        group.by = "sample")
```
```{r}
markers.to.plot <-
  c(
"Clu", 
"Gfap", 
"Aqp4", 
"Gja1", 
"Cldn5", 
"Adgrf5", 
"Flt1", 
"Col1a1", 
"Col1a2", 
"Dcn", 
"Hexb", 
"C1qa", 
"C1qc", 
"C1qb", 
"Tmem119", 
"Itgam", 
"Tyrobp", 
"P2ry12", 
"Aif1", 
"Rbfox3", 
"Snap25", 
"Syt1", 
"Gad1", 
"Gad2", 
"Plp1", 
"Mbp", 
"Mog", 
"Olig1", 
"Pdgfra", 
"Vcan", 
"Tnr", 
"Acta2", 
"Rgs5", 
"Vtn", 
"Myl5"
  )

dot_ind <- DotPlot(dataObject,
                   features = markers.to.plot, 
                   cluster.idents = TRUE,
                   dot.scale = 8) + RotatedAxis()
dot_ind
```


# Unannotated 
### UMAP
```{r unannotated_umap}
ditto_umap <- dittoDimPlot(object = dataObject,
             var = "seurat_clusters",
             do.label = TRUE,
             labels.highlight = TRUE)
ditto_umap
```

### Nuclei count per cluster
```{r count_per_cluster}
count_per_cluster <- FetchData(dataObject,
                               vars = c("ident", "orig.ident")) %>%
  dplyr::count(ident, orig.ident) %>%
  tidyr::spread(ident, n)
count_per_cluster

count_melt <- reshape2::melt(count_per_cluster)
colnames(count_melt) <- c("ident", "cluster", "number of nuclei")
count_max <- count_melt[which.max(count_melt$`number of nuclei`), ]
count_max_value <- count_max$`number of nuclei`
cellmax <- count_max_value + 200 # so that the figure doesn't cut off the text
count_bar <- ggplot(count_melt, aes(x = factor(cluster), y = `number of nuclei`, fill = `ident`)) +
  geom_bar(
    stat = "identity",
    colour = "black",
    width = 1,
    position = position_dodge(width = 0.8)
  ) +
  geom_text(
    aes(label = `number of nuclei`),
    position = position_dodge(width = 0.9),
    vjust = -0.25,
    angle = 45,
    hjust = -.01
  ) +
  theme_classic() + 
  #scale_fill_manual(values = sample_colors) +
  ggtitle("Number of nuclei per cluster") +  xlab("cluster") +
  theme(axis.text.x = element_text(angle = 45, hjust = 1)) +
  scale_y_continuous(limits = c(0, cellmax))
count_bar
```

### Cluster tree
```{r cluster_tree}
dataObject <- BuildClusterTree(
  object = dataObject,
  dims = 1:30,
  reorder = FALSE,
  reorder.numeric = FALSE
)

tree <- dataObject@tools$BuildClusterTree
tree$tip.label <- paste0("Cluster ", tree$tip.label)
nClusters <- length(tree$tip.label)

tree_graph <- ggtree::ggtree(tree, aes(x, y)) +
  scale_y_reverse() +
  ggtree::geom_tree() +
  ggtree::theme_tree() +
  ggtree::geom_tiplab(offset = 1) +
  ggtree::geom_tippoint(color = color.panel[1:nClusters],
                        shape = 16,
                        size = 5) +
  coord_cartesian(clip = 'off') +
  theme(plot.margin = unit(c(0, 2.5, 0, 0), 'cm'))
tree_graph
```

## Violins - Canonical cell-type markers
Markers obtained from dropviz
```{r violins}
# Neurons
VlnPlot(dataObject,
        features = c("Rbfox3", "Snap25", "Syt1", "Gad1", "Gad2"))
# Astrocytes
VlnPlot(dataObject,
        features = c("Aqp4", "Gfap", "Clu"))
# Microglia  
VlnPlot(dataObject,
        features = c("Hexb", "C1qa", "C1qc", "C1qb"))
## DAM
VlnPlot(dataObject,
        features = c("Apoe", "Cst7", "Lyz2", "Lpl", "Cd9", "Trem2"))

## Homeostatis
VlnPlot(dataObject,
        features = c("P2ry12", "Cx3cr1"))
# Oligodendrocyte
VlnPlot(dataObject,
        features = c("Plp1","Mbp", "Mog", "Mag"))
# OPC
VlnPlot(dataObject,
        features = c("Olig1","Pdgfra", "Vcan", "Tnr"))
# Microglia / Marchophage 
VlnPlot(dataObject,
        features = c( "F13a1", "Mrc1"))
# Endothelial
VlnPlot(dataObject,
        features = c("Igfbp7", "Fn1", "Sox17", "Cldn5", "Adgrf5", "Flt1"))
# Fibroblast-Like_Dcn
VlnPlot(dataObject,
        features = c("Col1a1", "Col1a2", "Dcn"))
# Smooth muscle 
VlnPlot(dataObject,
        features = c("Acta2", "Rgs5", "Vtn", "Myl5"))
```

## Percent cell type - Canonical cell-type markers
```{r percent_cell_type}
# neuron 
dataObject[["percent.neurons"]] <- PercentageFeatureSet(dataObject, 
                                   features = c("Rbfox3", "Snap25", "Syt1", "Gad1", "Gad2"))
# astrocyte
dataObject[["percent.astrocyte"]] <- PercentageFeatureSet(dataObject, 
                                     features = c("Aqp4", "Gfap", "Clu"))
# microglia
dataObject[["percent.microglia"]] <- PercentageFeatureSet(dataObject, 
                                     features = c("Hexb", "C1qa", "C1qc", "C1qb")) 
# DAM
dataObject[["percent.DAM"]] <- PercentageFeatureSet(dataObject, 
                                     features = c("Apoe", "Cst7", "Lyz2", "Lpl", "Cd9", "Trem2")) 
# homeostatis
dataObject[["percent.homeostatis"]] <- PercentageFeatureSet(dataObject, 
                                     features = c("P2ry12", "Cx3cr1")) 
# oligodendrocyte
dataObject[["percent.oligodendrocyte"]] <- PercentageFeatureSet(dataObject, 
                                           features = c("Plp1","Mbp", "Mog", "Mag"))
# oligodendrocyte precursor cells
dataObject[["percent.opc"]] <- PercentageFeatureSet(dataObject, 
                               features = c("Olig1","Pdgfra", "Vcan", "Tnr"))
# Marchophage
dataObject[["percent.marchophage"]] <- PercentageFeatureSet(dataObject, 
                               features = c("F13a1", "Mrc1"))
# endothelial
dataObject[["percent.endothelial"]] <- PercentageFeatureSet(dataObject, 
                                       features = c("Igfbp7", "Fn1", "Sox17", "Cldn5", "Adgrf5", "Flt1"))
# fibroblast 
dataObject[["percent.fibroblast"]] <- PercentageFeatureSet(dataObject, 
                                      features = c("Col1a1", "Col1a2", "Dcn"))
# smooth muscle 
dataObject[["percent.smooth"]] <- PercentageFeatureSet(dataObject, 
                                  features = c("Acta2", "Rgs5", "Vtn"))
```

## Feature plot percent cell type - Canonical cell-type markers
```{r featureplot_canonical_markers}
# astrocyte
FeaturePlot(dataObject, features = "percent.astrocyte", label = TRUE)  
# endothelial
FeaturePlot(dataObject, features = "percent.endothelial", label = TRUE) 
# fibroblast
FeaturePlot(dataObject, features = "percent.fibroblast", label = TRUE) 
# microglia
FeaturePlot(dataObject, features = "percent.microglia", label = TRUE)  
FeaturePlot(dataObject, features = "percent.DAM", label = TRUE)  
FeaturePlot(dataObject, features = "percent.homeostatis", label = TRUE)  

# neuron 
FeaturePlot(dataObject, features = "percent.neurons", label = TRUE) 
# oligodendrocyte
FeaturePlot(dataObject, features = "percent.oligodendrocyte", label = TRUE) 
# oligodendrocyte precursor cells
FeaturePlot(dataObject, features = "percent.opc", label = TRUE)  
# smooth
FeaturePlot(dataObject, features = "percent.smooth", label = TRUE) 
```

## Markers per cluster
```{r find_all_markers}
markers <- SeuratWrappers::RunPrestoAll(
  object = dataObject,
  assay = "RNA",
  slot = "counts",
  only.pos = FALSE
)
write.table(markers, 
            paste0("../results/markers/", treatment, "_markers.tsv"),
            quote = FALSE,
            row.names = FALSE)
saveRDS(markers, paste0("../rObjects/", treatment, "_markers.rds"))

# rearrange to order by cluster & filter to only include log2FC > 1 & FDR < 0.05
 all.markers.strict <- markers %>%
   group_by(cluster) %>%
   dplyr::filter(avg_log2FC > 1 & p_val_adj < 0.05)

saveRDS(all.markers.strict, paste0("../rObjects/", treatment,"_markers_log2FC1_q0.01.rds"))
```


# Get markers for each cluster
```{r markers_per_cluster}
# unique clusters variable
unique_clusters <- unique(markers$cluster)

# empty list to store individual cluster data frames
cluster_list <- list()

# loop through each cluster and create a data frame
for (i in unique_clusters) {
  cluster_name <- paste0("cluster", i)
  cluster_data <- all.markers.strict[all.markers.strict$cluster == i, ]
  assign(cluster_name, cluster_data)
  cluster_list[[cluster_name]] <- cluster_data
}
```

## Feature plot
```{r featureplot}
# UMAP showing the expression of select features
umap_feature <-
  FeaturePlot(dataObject,
              features = c("Tyrobp", "Mog", "Aqp4", "Rbfox3"))
umap_feature
```
```{r}
markers.to.plot <-
  c(
"Clu", 
"Gfap", 
"Aqp4", 
"Gja1", 
"Cldn5", 
"Adgrf5", 
"Flt1", 
"Col1a1", 
"Col1a2", 
"Dcn", 
"Hexb", 
"C1qa", 
"C1qc", 
"C1qb", 
"Tmem119", 
"Itgam", 
"Tyrobp", 
"P2ry12", 
"Aif1", 
"Rbfox3", 
"Snap25", 
"Syt1", 
"Gad1", 
"Gad2", 
"Plp1", 
"Mbp", 
"Mog", 
"Olig1", 
"Pdgfra", 
"Vcan", 
"Tnr", 
"Acta2", 
"Rgs5", 
"Vtn", 
"Myl5"
  )

dot_ind <- DotPlot(dataObject,
                   features = markers.to.plot, 
                   cluster.idents = FALSE,
                   dot.scale = 8) + RotatedAxis()
dot_ind

help(DotPlot)
```

```{r save_dot_individual, echo=FALSE}
pdf(
  paste0(
    "../results/dot_plot/",
    treatment,
    "_clusters_DotPlot_no_integration_order_by_cluster.pdf"
  ),
  width = 14,
  height = 9
)
dot_ind
dev.off()
```

# Cluster Annotation

## Cluster 0 - interneuron
```{r cluster0}
# Number of cells per condition
count_per_cluster[,c(1,2)]
# UMAP with only cluster 0
DimPlot(object = subset(dataObject, seurat_clusters == "0"),
        reduction = "umap", 
        label = TRUE,
        label.box = TRUE,
        label.size = 3,
        repel = TRUE,
        cols = color.panel[1])
VlnPlot(dataObject,
        features = cluster0$gene[1:10],
        stack = TRUE,
        flip = TRUE,
        split.by = "seurat_clusters")
cluster0$gene[1:10]
```

## Cluster 1 - neuron
```{r cluster1}
count_per_cluster[,c(1,3)]
DimPlot(object = subset(dataObject, seurat_clusters == "1"),
        reduction = "umap", 
        label = TRUE,
        label.box = TRUE,
        label.size = 3,
        repel = TRUE,
        cols = color.panel[2])
VlnPlot(dataObject,
        features = cluster1$gene[1:10],
        cols = color.panel,
        stack = TRUE,
        flip = TRUE,
        split.by = "seurat_clusters")
cluster1$gene[1:10]
```

## Cluster 2 - interneuron
```{r cluster2}
count_per_cluster[,c(1,4)]
DimPlot(object = subset(dataObject, seurat_clusters == "2"),
        reduction = "umap", 
        label = TRUE,
        label.box = TRUE,
        label.size = 3,
        repel = TRUE,
        cols = color.panel[3])
VlnPlot(dataObject,
        features = cluster2$gene[1:10],
        cols = color.panel,
        stack = TRUE,
        flip = TRUE,
        split.by = "seurat_clusters")
cluster2$gene[1:10]
```

## Cluster 3 - neuron 
```{r cluster3}
count_per_cluster[,c(1,4)]
DimPlot(object = subset(dataObject, seurat_clusters == "3"),
        reduction = "umap", 
        label = TRUE,
        label.box = TRUE,
        label.size = 3,
        repel = TRUE,
        cols =  color.panel[4])
VlnPlot(dataObject,
        features = cluster3$gene[1:10],
        cols = color.panel,
        stack = TRUE,
        flip = TRUE,
        split.by = "seurat_clusters")
cluster3$gene[1:10]
```

## Cluster 4 - astrocyte
```{r cluster4}
count_per_cluster[,c(1,5)]
DimPlot(object = subset(dataObject, seurat_clusters == "4"),
        reduction = "umap", 
        label = TRUE,
        label.box = TRUE,
        label.size = 3,
        repel = TRUE,
        cols = color.panel[5])
VlnPlot(dataObject,
        features = cluster4$gene[1:10],
        cols = color.panel,
        stack = TRUE,
        flip = TRUE,
        split.by = "seurat_clusters")
cluster4$gene[1:10]
```

## Cluster 5 - neuron
```{r cluster5}
count_per_cluster[,c(1,6)]
DimPlot(object = subset(dataObject, seurat_clusters == "5"),
        reduction = "umap", 
        label = TRUE,
        label.box = TRUE,
        label.size = 3,
        repel = TRUE,
        cols = color.panel[6])
VlnPlot(dataObject,
        features = cluster5$gene[1:10],
        cols = color.panel,
        stack = TRUE,
        flip = TRUE,
        split.by = "seurat_clusters")
cluster5$gene[1:10]
```

## Cluster 6 - neuron
```{r cluster6}
count_per_cluster[,c(1,7)]
DimPlot(object = subset(dataObject, seurat_clusters == "6"),
        reduction = "umap", 
        label = TRUE,
        label.box = TRUE,
        label.size = 3,
        repel = TRUE,
        cols = "#CC79A7")
VlnPlot(dataObject,
        features = cluster6$gene[1:10],
        cols = color.panel,
        stack = TRUE,
        flip = TRUE,
        split.by = "seurat_clusters")
cluster6$gene[1:10]
```

## Cluster 7 - interneuron
```{r cluster7}
count_per_cluster[,c(1,8)]
DimPlot(object = subset(dataObject, seurat_clusters == "7"),
        reduction = "umap", 
        label = TRUE,
        label.box = TRUE,
        label.size = 3,
        repel = TRUE,
        cols = color.panel[7])
VlnPlot(dataObject,
        features = cluster7$gene[1:10],
        cols = color.panel,
        stack = TRUE,
        flip = TRUE,
        split.by = "seurat_clusters")
cluster7$gene[1:10]
```

## Cluster 8 - oligodendrocyte
```{r cluster8}
count_per_cluster[,c(1,9)]
DimPlot(object = subset(dataObject, seurat_clusters == "8"),
        reduction = "umap", 
        label = TRUE,
        label.box = TRUE,
        label.size = 3,
        repel = TRUE,
        cols = color.panel[8])
VlnPlot(dataObject,
        features = cluster8$gene[1:10],
        cols = color.panel,
        stack = TRUE,
        flip = TRUE,
        split.by = "seurat_clusters")
cluster8$gene[1:10]
```

## Cluster 9 - neuron
```{r cluster9}
count_per_cluster[,c(1,10)]
DimPlot(object = subset(dataObject, seurat_clusters == "9"),
        reduction = "umap", 
        label = TRUE,
        label.box = TRUE,
        label.size = 3,
        repel = TRUE,
        cols = color.panel[9])
VlnPlot(dataObject,
        features = cluster9$gene[1:10],
        cols = color.panel,
        stack = TRUE,
        flip = TRUE,
        split.by = "seurat_clusters")
cluster9$gene[1:10]
```

## Cluster 10 - astrocyte
```{r cluster10}
count_per_cluster[,c(1,11)]
DimPlot(object = subset(dataObject, seurat_clusters == "10"),
        reduction = "umap", 
        label = TRUE,
        label.box = TRUE,
        label.size = 3,
        repel = TRUE,
        cols = color.panel[10])
VlnPlot(dataObject,
        features = cluster10$gene[1:10],
        cols = color.panel,
        stack = TRUE,
        flip = TRUE,
        split.by = "seurat_clusters")
cluster10$gene[1:10]
```

## Cluster 11 - oligodendrocyte with noise 
```{r cluster11}
count_per_cluster[,c(1,12)]
DimPlot(object = subset(dataObject, seurat_clusters == "11"),
        reduction = "umap", 
        label = TRUE,
        label.box = TRUE,
        label.size = 3,
        repel = TRUE,
        cols = color.panel[11])
VlnPlot(dataObject,
        features = cluster11$gene[1:10],
        cols = color.panel,
        stack = TRUE,
        flip = TRUE,
        split.by = "seurat_clusters")
cluster11$gene[1:10]
```

## Cluster 12 - neuron
```{r cluster12}
count_per_cluster[,c(1,13)]
DimPlot(object = subset(dataObject, seurat_clusters == "12"),
        reduction = "umap", 
        label = TRUE,
        label.box = TRUE,
        label.size = 3,
        repel = TRUE,
        cols = color.panel[12])
VlnPlot(dataObject,
        features = cluster12$gene[1:10],
        cols = color.panel,
        stack = TRUE,
        flip = TRUE,
        split.by = "seurat_clusters")
cluster12$gene[1:10]
```

## Cluster 13 - polydendrocyte
```{r cluster13}
count_per_cluster[,c(1,14)]
DimPlot(object = subset(dataObject, seurat_clusters == "13"),
        reduction = "umap", 
        label = TRUE,
        label.box = TRUE,
        label.size = 3,
        repel = TRUE,
        cols = color.panel[13])
VlnPlot(dataObject,
        features = cluster13$gene[1:10],
        cols = color.panel,
        stack = TRUE,
        flip = TRUE,
        split.by = "seurat_clusters")
cluster13$gene[1:10]
```

## Cluster 14 - interneuron
```{r cluster14}
count_per_cluster[,c(1,15)]
DimPlot(object = subset(dataObject, seurat_clusters == "14"),
        reduction = "umap", 
        label = TRUE,
        label.box = TRUE,
        label.size = 3,
        repel = TRUE,
        cols = color.panel[14])
VlnPlot(dataObject,
        features = cluster14$gene[1:10],
        cols = color.panel,
        stack = TRUE,
        flip = TRUE,
        split.by = "seurat_clusters")
cluster14$gene[1:10]
```

## Cluster 15 - neuron
```{r cluster15}
count_per_cluster[,c(1,16)]
DimPlot(object = subset(dataObject, seurat_clusters == "15"),
        reduction = "umap", 
        label = TRUE,
        label.box = TRUE,
        label.size = 3,
        repel = TRUE,
        cols = color.panel[15])
VlnPlot(dataObject,
        features = cluster15$gene[1:10],
        cols = color.panel,
        stack = TRUE,
        flip = TRUE,
        split.by = "seurat_clusters")
cluster15$gene[1:10]
```
## Cluster 16 - fibroblast
```{r cluster16}
count_per_cluster[,c(1,17)]
DimPlot(object = subset(dataObject, seurat_clusters == "16"),
        reduction = "umap", 
        label = TRUE,
        label.box = TRUE,
        label.size = 3,
        repel = TRUE,
        cols = color.panel[16])
VlnPlot(dataObject,
        features = cluster16$gene[1:10],
        cols = color.panel,
        stack = TRUE,
        flip = TRUE,
        split.by = "seurat_clusters")
cluster16$gene[1:10]
```

## Cluster 17 - nueron
```{r cluster17}
count_per_cluster[,c(1,18)]
DimPlot(object = subset(dataObject, seurat_clusters == "17"),
        reduction = "umap", 
        label = TRUE,
        label.box = TRUE,
        label.size = 3,
        repel = TRUE,
        cols = color.panel[17])
VlnPlot(dataObject,
        features = cluster17$gene[1:10],
        cols = color.panel,
        stack = TRUE,
        flip = TRUE,
        split.by = "seurat_clusters")
cluster17$gene[1:10]
```

## Cluster 18 - astrocyte - noise 
```{r cluster18}
count_per_cluster[,c(1,1)]
DimPlot(object = subset(dataObject, seurat_clusters == "18"),
        reduction = "umap", 
        label = TRUE,
        label.box = TRUE,
        label.size = 3,
        repel = TRUE,
        cols = color.panel[19])
VlnPlot(dataObject,
        features = cluster18$gene[1:10],
        cols = color.panel,
        stack = TRUE,
        flip = TRUE,
        split.by = "seurat_clusters")
cluster18$gene[1:10]
```

## Cluster 19 - fibroblast
```{r cluster19}
count_per_cluster[,c(1,19)]
DimPlot(object = subset(dataObject, seurat_clusters == "19"),
        reduction = "umap", 
        label = TRUE,
        label.box = TRUE,
        label.size = 3,
        repel = TRUE,
        cols = color.panel[20])
VlnPlot(dataObject,
        features = cluster19$gene[1:10],
        cols = color.panel,
        stack = TRUE,
        flip = TRUE,
        split.by = "seurat_clusters")
cluster19$gene[1:10]
```


## Cluster 20 - interneuron
```{r cluster20}
count_per_cluster[,c(1,20)]
DimPlot(object = subset(dataObject, seurat_clusters == "20"),
        reduction = "umap", 
        label = TRUE,
        label.box = TRUE,
        label.size = 3,
        repel = TRUE,
        cols = color.panel[21])
VlnPlot(dataObject,
        features = cluster20$gene[1:10],
        cols = color.panel,
        stack = TRUE,
        flip = TRUE,
        split.by = "seurat_clusters")
cluster20$gene[1:10]
```

## Cluster 21 - endothelial
```{r cluster21}
count_per_cluster[,c(1,21)]
DimPlot(object = subset(dataObject, seurat_clusters == "21"),
        reduction = "umap", 
        label = TRUE,
        label.box = TRUE,
        label.size = 3,
        repel = TRUE,
        cols = color.panel[22])
VlnPlot(dataObject,
        features = cluster21$gene[1:10],
        cols = color.panel,
        stack = TRUE,
        flip = TRUE,
        split.by = "seurat_clusters")
cluster21$gene[1:10]
```

## Cluster 22 - interneuron
```{r cluster22}
count_per_cluster[,c(1,22)]
DimPlot(object = subset(dataObject, seurat_clusters == "22"),
        reduction = "umap", 
        label = TRUE,
        label.box = TRUE,
        label.size = 3,
        repel = TRUE,
        cols = color.panel[23])
VlnPlot(dataObject,
        features = cluster22$gene[1:10],
        cols = color.panel,
        stack = TRUE,
        flip = TRUE,
        split.by = "seurat_clusters")
cluster22$gene[1:10]
```

## Cluster 23 - neuron
```{r cluster23}
count_per_cluster[,c(1,23)]
DimPlot(object = subset(dataObject, seurat_clusters == "23"),
        reduction = "umap", 
        label = TRUE,
        label.box = TRUE,
        label.size = 3,
        repel = TRUE,
        cols = color.panel[24])
VlnPlot(dataObject,
        features = cluster23$gene[1:10],
        cols = color.panel,
        stack = TRUE,
        flip = TRUE,
        split.by = "seurat_clusters")
cluster23$gene[1:10]
```

## Cluster 24 - neuron
```{r cluster24}
count_per_cluster[,c(1,24)]
DimPlot(object = subset(dataObject, seurat_clusters == "24"),
        reduction = "umap", 
        label = TRUE,
        label.box = TRUE,
        label.size = 3,
        repel = TRUE,
        cols = color.panel[25])
VlnPlot(dataObject,
        features = cluster24$gene[1:10],
        cols = color.panel,
        stack = TRUE,
        flip = TRUE,
        split.by = "seurat_clusters")
cluster24$gene[1:10]
```

## Cluster 25 - neuron
```{r cluster25}
count_per_cluster[,c(1,25)]
DimPlot(object = subset(dataObject, seurat_clusters == "25"),
        reduction = "umap", 
        label = TRUE,
        label.box = TRUE,
        label.size = 3,
        repel = TRUE,
        cols = color.panel[26])
VlnPlot(dataObject,
        features = cluster25$gene[1:10],
        cols = color.panel,
        stack = TRUE,
        flip = TRUE,
        split.by = "seurat_clusters")
cluster25$gene[1:10]
```

## Cluster 26 - mural
```{r cluster26}
count_per_cluster[,c(1,26)]
DimPlot(object = subset(dataObject, seurat_clusters == "26"),
        reduction = "umap", 
        label = TRUE,
        label.box = TRUE,
        label.size = 3,
        repel = TRUE,
        cols = color.panel[27])
VlnPlot(dataObject,
        features = cluster26$gene[1:10],
        cols = color.panel,
        stack = TRUE,
        flip = TRUE,
        split.by = "seurat_clusters")
cluster26$gene[1:10]
```

## Cluster 27 - microglia
```{r cluster27}
count_per_cluster[,c(1,27)]
DimPlot(object = subset(dataObject, seurat_clusters == "27"),
        reduction = "umap", 
        label = TRUE,
        label.box = TRUE,
        label.size = 3,
        repel = TRUE,
        cols = color.panel[28])
VlnPlot(dataObject,
        features = cluster27$gene[1:10],
        cols = color.panel,
        stack = TRUE,
        flip = TRUE,
        split.by = "seurat_clusters")
cluster27$gene[1:10]
```

## Cluster 28 - neuron
```{r cluster28}
count_per_cluster[,c(1,28)]
DimPlot(object = subset(dataObject, seurat_clusters == "28"),
        reduction = "umap", 
        label = TRUE,
        label.box = TRUE,
        label.size = 3,
        repel = TRUE,
        cols = color.panel[28])
VlnPlot(dataObject,
        features = cluster28$gene[1:10],
        cols = color.panel,
        stack = TRUE,
        flip = TRUE,
        split.by = "seurat_clusters")
cluster28$gene[1:10]
```

# Assign identities
## Individual 
```{r assign_individual}
dataObject.annotated <- RenameIdents(object = dataObject, 
                               "0" = "interneuron",
                               "1" = "neuron",
                               "2" = "interneuron",
                               "3" = "neuron",
                               "4" = "astrocyte",
                               "5" = "neuron",
                               "6" = "neuron",
                               "7" = "interneuron",
                               "8" = "oligodendrocyte",
                               "9" = "neuron",
                               "10" = "astrocyte",
                               "11" = "oligodendrocyte",
                               "12" = "neuron",
                               "13" = "polydendrocyte",
                               "14" = "interneuron",
                               "15" = "neuron",
                               "16" = "fibroblast",
                               "17" = "neuron",
                               "18" = "astrocyte",
                               "19" = "fibroblast",
                               "20" = "interneuron",
                               "21" = "endothelial",
                               "22" = "interneuron",
                               "23" = "neuron",
                               "24" = "neuron",
                               "25" = "neuron",
                               "26" = "mural",
                               "27" = "microglia",
                               "28" = "neuron"                              )
dataObject.annotated$individual_clusters <- factor(Idents(dataObject.annotated))

UMAP_ind <- dittoDimPlot(object = dataObject.annotated,
             var = "individual_clusters",
             reduction.use = "umap",
             do.label = TRUE,
             labels.highlight = TRUE)
UMAP_ind
```

```{r UMAP_ind, echo=FALSE}
pdf(
  paste0(
    "../results/UMAP/annotated/",
    treatment,
    "_individual_clusters_UMAP.pdf"
  ),
  width = 11,
  height = 7
)
UMAP_ind
dev.off()
```

### Nuclei count per cluster
```{r annotated_counts_individual}
count_per_cluster <- FetchData(dataObject.annotated,
                               vars = c("ident", "orig.ident")) %>%
  dplyr::count(ident, orig.ident) %>%
  tidyr::spread(ident, n)
count_per_cluster

count_melt <- reshape2::melt(count_per_cluster)
colnames(count_melt) <- c("ident", "cluster", "number of nuclei")
count_max <- count_melt[which.max(count_melt$`number of nuclei`), ]
count_max_value <- count_max$`number of nuclei`
cellmax <- count_max_value + 500 # so that the figure doesn't cut off the text
count_bar <- ggplot(count_melt, aes(x = factor(cluster), y = `number of nuclei`, fill = `cluster`)) +
  geom_bar(
    stat = "identity",
    colour = "black",
    width = 1,
    position = position_dodge(width = 0.8)
  ) +
  geom_text(
    aes(label = `number of nuclei`),
    position = position_dodge(width = 0.9),
    vjust = -0.25,
    angle = 45,
    hjust = -.01
  ) +
  theme_classic() + scale_fill_manual(values = color.panel) +
  ggtitle("Number of nuclei per cluster") +  xlab("cluster") +
  theme(axis.text.x = element_text(angle = 45, hjust = 1)) +
  scale_y_continuous(limits = c(0, cellmax))
count_bar
```

```{r nuclei_count_individual, echo=FALSE}
pdf(
  paste0(
    "../results/UMAP/annotated/",
    treatment,
    "_individual_clusters_nuclei_count.pdf"
  ),
  width = 11,
  height = 7
)
count_bar
dev.off()
```

### DotPlot
```{r dot_individual}
markers.to.plot <-
  c(
"Clu", 
"Gfap", 
"Aqp4", 
"Gja1", 
"Cldn5", 
"Adgrf5", 
"Flt1", 
"Col1a1", 
"Col1a2", 
"Dcn", 
"Hexb", 
"C1qa", 
"C1qc", 
"C1qb", 
"Tmem119", 
"Itgam", 
"Tyrobp", 
"P2ry12", 
"Aif1", 
"Rbfox3", 
"Snap25", 
"Syt1", 
"Gad1", 
"Gad2", 
"Plp1", 
"Mbp", 
"Mog", 
"Olig1", 
"Pdgfra", 
"Vcan", 
"Tnr", 
"Acta2", 
"Rgs5", 
"Vtn", 
"Myl5"
  )

dot_ind <- DotPlot(dataObject.annotated,
                   features = markers.to.plot, 
                   split.by = "strain", 
                   cluster.idents = TRUE,
                   dot.scale = 8) + RotatedAxis()
dot_ind
```

```{r save_dot_individual, echo=FALSE}
pdf(
  paste0(
    "../results/dot_plot/",
    treatment,
    "_clusters_DotPlot_no_integration_split_strain.pdf"
  ),
  width = 14,
  height = 5
)
dot_ind
dev.off()
```


### Relative abundance
```{r relative_cell_type}
relative_abundance <- dataObject.annotated@meta.data %>%
  group_by(individual_clusters, orig.ident) %>%
  dplyr::count() %>%
  group_by(orig.ident) %>%
  dplyr::mutate(percent = 100 * n / sum(n)) %>%
  ungroup()


rel_abun <- ggplot(relative_abundance, aes(x = orig.ident, y = percent, fill = individual_clusters)) +
  geom_col() +
  geom_text(aes(label = paste0(round(percent), "%")), 
            position = position_stack(vjust = 0.5), size = 3, color = "white") +
  scale_fill_manual(values = color.panel) +
  ggtitle("Percentage of cell type") +
  theme_bw() +
  theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust = 1))
rel_abun

pdf(
  paste0(
    "../results/UMAP/annotated/",
    treatment,
    "_merged_clusters_relative_abundance.pdf"
  ),
  width = 6,
  height = 8
)
rel_abun
dev.off()
```


# Save RDS 
```{r save_annotations}
saveRDS(dataObject.annotated, paste0("../rObjects/", treatment, ".annotated.rds"))
```

```{r}
```
