==> scripts/13_correlation_comparisons_among_DE_tools.Rmd <==
---
title: "Correlation in the log2FC comparisons"
author: "Kimberly Olney, PhD"
date: "05/29/2024"
output:
  html_document:
    theme: cerulean
    toc: true
    toc_float: true
editor_options: 
  chunk_output_type: inline
---

Differential expression was completed for the following comparisons:
1.	PS19 vs WT within each cell type. All samples.\
2.	PS19 vs WT within each cell type. 0 hours only.\
3.	PS19 vs WT within each cell type. 3 hours only.\

```{r setup, include=FALSE}
knitr::opts_chunk$set(root.dir = ".", echo=TRUE, warning=FALSE, message=FALSE)
library(ComplexHeatmap)
library(ggpubr)
```

# Libraris, paths, colors
https://satijalab.org/seurat/articles/essential_commands.html 
```{r echo=FALSE, message=FALSE}
source(here::here("/research/labs/neurology/fryer/m239830/PMI/scripts", "file_paths_and_colours.R"))
treatment <- "PMI"
color.panel <- dittoColors()
metadata <- subset(metadata, sampleID != "P_60" & sampleID != "B_6")
metadata$sampleID <- factor(metadata$sampleID, levels = c(metadata$sampleID))
samples <- metadata$sampleID 
```

# Read in object
```{r read_object}
# read object
dataObject <- readRDS(file = paste0("../rObjects/", treatment, ".dataObject.clean.rds"))
# inspect
dataObject
Layers(dataObject)
# List of cell types to analyze
cell_types <- c(unique(dataObject$individual_clusters))
```

```{r protein_coding}
protein_coding_genes <- subset(gene_info, type == "protein_coding")
```

# Sex check
```{r exp_matrix}
# Pseudo-bulk the counts based on sample, strain, sex, time, and cell type
dataObject.pseudo <- AggregateExpression(
    dataObject, 
    assays = "RNA", # DESeq works with raw counts
    return.seurat = TRUE, 
    features = protein_coding_genes$gene_name,
    group.by = c("sample", "strain", "sex", "time", "individual_clusters")
)
# Normalize the data
dataObject.pseudo <- NormalizeData(dataObject.pseudo, normalization.method = "LogNormalize", scale.factor = 10000)
# Find variable features
dataObject.pseudo <- FindVariableFeatures(dataObject.pseudo, selection.method = "vst", nfeatures = 2000)

# Scale the data
dataObject.pseudo <- ScaleData(dataObject.pseudo)

# Perform PCA
dataObject.pseudo <- RunPCA(dataObject.pseudo, features = VariableFeatures(object = dataObject.pseudo))

# Visualize the expression of a specific gene
FeaturePlot(dataObject.pseudo, features = "Rbfox3")

pseudo_bulk_counts <- GetAssayData(dataObject.pseudo, layer = "counts")
pseudo_bulk_counts <- NormalizeData(pseudo_bulk_counts)
df_norm <- as.matrix(pseudo_bulk_counts)
```

# Expression heatmap for each cell type, all expressed genes 
```{r heatmap_celltype, eval=FALSE}
# Define colors for each level of qualitative variables
col = list(
  strain = c("PS19" = "orange", "WT" = "yellow"),
  sex = c("Male" = "green", "Female" = "purple"),
  timepoint = c("3" = "black", "0" = "gray")
)

# Loop over each cell type
for (cell_type in cell_types) {
  # Filter the data for the current cell type
  df_filtered <- df_norm[, grep(paste0("_", cell_type, "$"), colnames(df_norm))]
  
   # Filter genes with counts above 1 in at least 3 samples
  gene_filter <- rowSums(df_filtered > 1) >= 3
  df_filtered <- df_filtered[gene_filter, ]
  
  # Extract annotations from column names
  annotations <- data.frame(
    strain = sapply(strsplit(colnames(df_filtered), "_"), function(x) x[2]), 
    sex = sapply(strsplit(colnames(df_filtered), "_"), function(x) x[3]),
    timepoint = sapply(strsplit(colnames(df_filtered), "_"), function(x) x[4])
  )
  
  # Create the heatmap annotation
  ha <- HeatmapAnnotation(
    df = annotations,
    col = col
  )
  
  # Create the heatmap
  heatmap_plot <- Heatmap(df_filtered, name = "Expression",
                          top_annotation = ha,
                          show_row_names = FALSE)
  
  # Save the heatmap as a PDF
  pdf(paste0("../results/heatmap/", cell_type, "_expression.pdf"),
      width = 8.5,
      height = 10)
  draw(heatmap_plot)
  dev.off()
}
```

# 3 vs 0 hours DEGs heatmap
```{r heatmap_3_vs_0}
# Define colors for each level of qualitative variables
col = list(
  strain = c("PS19" = "orange", "WT" = "yellow"),
  sex = c("Male" = "green", "Female" = "purple"),
  timepoint = c("3" = "black", "0" = "gray")
)

# Loop over each cell type
for (cell_type in cell_types) {
  # Filter the data for the current cell type
  df_filtered <- df_norm[, grep(paste0("_", cell_type, "$"), colnames(df_norm))]

  DEGs <- read.delim(paste0("../results/DEGs/DESeq2_pseudobulk_exp_filter/", cell_type,
                            "_3_vs_0_hours_comparison_pseudobulk.txt"))
  gene_filter <- subset(DEGs, (padj < 0.1 & log2FoldChange > 0) | ((padj < 0.1 & log2FoldChange < -0)))
  df_filtered <- df_filtered[gene_filter$gene, ]
  
  # Extract annotations from column names
  annotations <- data.frame(
    strain = sapply(strsplit(colnames(df_filtered), "_"), function(x) x[2]), 
    sex = sapply(strsplit(colnames(df_filtered), "_"), function(x) x[3]),
    timepoint = sapply(strsplit(colnames(df_filtered), "_"), function(x) x[4])
  )
  
  # Create the heatmap annotation
  ha <- HeatmapAnnotation(
    df = annotations,
    col = col
  )
  
  # Create the heatmap
  heatmap_plot <- Heatmap(df_filtered, name = "Expression",
                          top_annotation = ha,
                          show_row_names = TRUE, # show gene names
                          column_names_rot = 75)  # Rotate column names

  
  # Save the heatmap as a PDF
  pdf(paste0("../results/heatmap/", cell_type, "_expression_3_vs_0_hours_DEGs_q0.1.pdf"),
      width = 14,
      height = 10)
  draw(heatmap_plot)
  dev.off()
}
```

# PS19 vs WT  DEGs heatmap
```{r heatmap_PS19_vs_WT}
# Define colors for each level of qualitative variables
col = list(
  strain = c("PS19" = "orange", "WT" = "yellow"),
  sex = c("Male" = "green", "Female" = "purple"),
  timepoint = c("3" = "black", "0" = "gray")
)
   
cell_types <- c("neuron", "interneuron", "astrocyte", "polydendrocyte", "fibroblast", "mural", "microglia")
# Loop over each cell type
for (cell_type in cell_types) {
  # Filter the data for the current cell type
  df_filtered <- df_norm[, grep(paste0("_", cell_type, "$"), colnames(df_norm))]

  DEGs <- read.delim(paste0("../results/DEGs/DESeq2_pseudobulk_exp_filter/", cell_type,
                            "_PS19_vs_WT_comparison_pseudobulk.txt"))
  gene_filter <- subset(DEGs, (padj < 0.1 & log2FoldChange > 0.0) | ((padj < 0.1 & log2FoldChange < -0)))
  df_filtered <- df_filtered[gene_filter$gene, ]
  
  # Extract annotations from column names
  annotations <- data.frame(
    strain = sapply(strsplit(colnames(df_filtered), "_"), function(x) x[2]), 
    sex = sapply(strsplit(colnames(df_filtered), "_"), function(x) x[3]),
    timepoint = sapply(strsplit(colnames(df_filtered), "_"), function(x) x[4])
  )
  
  # Create the heatmap annotation
  ha <- HeatmapAnnotation(
    df = annotations,
    col = col
  )
  
  # Create the heatmap
  heatmap_plot <- Heatmap(df_filtered, name = "Expression",
                          top_annotation = ha,
                          show_row_names = TRUE, # show gene names
                          column_names_rot = 75)  # Rotate column names

  
  # Save the heatmap as a PDF
  pdf(paste0("../results/heatmap/", cell_type, "_expression_PS19_vs_WT_DEGs_q0.1.pdf"),
      width = 14,
      height = 10)
  draw(heatmap_plot)
  dev.off()
}
```

# Correlation matrix
```{r}
# Load necessary libraries
library(dplyr)

# Define the list of cell types and time points
time_points <- c("0_hours", "3_hours")

# Initialize an empty list to store data frames
deg_list <- list()

# Loop through each cell type and time point to read the data
for (cell_type in cell_types) {
  for (time_point in time_points) {
    # Construct the file name
    file_name <- paste0("../results/DEGs/DESeq2_pseudobulk_exp_filter/", cell_type, "_PS19_vs_WT_comparison_", time_point, "_only.txt")
    
    # Read the data
    deg_data <- read.table(file_name, header = TRUE, sep = "\t")
    
    # Add columns for cell type and time point
    deg_data <- deg_data %>%
      mutate(cell_type = cell_type, time_point = time_point)
    
    # Store the data frame in the list
    deg_list[[paste0(cell_type, "_", time_point)]] <- deg_data
  }
}

# Combine all data frames into a single data frame
deg_combined <- bind_rows(deg_list)

# Filter the data for 0 hours
deg_0hours <- deg_combined %>%
  filter(time_point == "0_hours") %>%
  select(gene, log2FoldChange, cell_type)

# Filter the data for 0 hours
deg_3hours <- deg_combined %>%
  filter(time_point == "3_hours") %>%
  select(gene, log2FoldChange, cell_type)

DEG_combined <- merge(deg_0hours, deg_3hours, by = c("gene", "cell_type"), suffixes = c("_0h", "_3h"))
head(DEG_combined)
```


# Plot the data
```{r}
DEG_combined2 <- na.omit(DEG_combined)

subset(DEG_combined2, gene == "Ctsb")
ggplot(DEG_combined2, aes(x = log2FoldChange_0h, y = log2FoldChange_3h)) +
  geom_point() +
  geom_smooth(method = "lm", se = FALSE, color = "blue") +
  facet_wrap(~ cell_type, ncol = 3) +
  stat_cor(method = "pearson", label.x = -4, label.y = 5) +
  labs(
    title = "Correlation of 0 hours and 3 hours by cell type",
    x = "0 hours log2FC",
    y = "3 hours log2FC"
  ) +
  theme_minimal()
```
# Correlation matrix MAST vs DESeq2
```{r}
# Load necessary libraries
library(dplyr)

DES_deg_list <- list()
for (cell_type in cell_types) {
    DES <- paste0("../results/DEGs/DESeq2_pseudobulk_exp_filter/", cell_type, "_PS19_vs_WT_comparison_pseudobulk.txt")
    DES_deg_data <- read.table(DES, header = TRUE, sep = "\t")
    DES_deg_data <- DES_deg_data %>%
      mutate(cell_type = cell_type)
    DES_deg_list[[paste0(cell_type, "_", "DESeq")]] <- DES_deg_data

}
DES_deg_combined <- bind_rows(DES_deg_list)


MAST_deg_list <- list()
for (cell_type in cell_types) {
    MAST <- paste0("../results/DEGs/MAST/", cell_type, "_PS19_vs_WT_comparison.txt")
    MAST_deg_data <- read.table(MAST, header = TRUE, sep = "\t")
    MAST_deg_data <- MAST_deg_data %>%
      mutate(cell_type = cell_type)
    MAST_deg_list[[paste0(cell_type, "_", "MAST")]] <- MAST_deg_data
}
MAST_deg_combined <- bind_rows(MAST_deg_list)

DEG_combined <- merge(DES_deg_combined, MAST_deg_combined, by = c("gene", "cell_type"), suffixes = c("DESeq", "MAST"))
head(DEG_combined)
```


# Plot the data
```{r}
DEG_combined2 <- na.omit(DEG_combined)

DEG_combined2$cell_type <- factor(DEG_combined2$cell_type, levels = c("neuron", "interneuron", "astrocyte", "oligodendrocyte", "fibroblast", "polydendrocyte", "endothelial", "mural", "microglia"))

ggplot(DEG_combined2, aes(x = log2FoldChange, y = avg_log2FC)) +
  geom_point() +
  geom_smooth(method = "lm", se = FALSE, color = "blue") +
  facet_wrap(~ cell_type, ncol = 3) +
  stat_cor(method = "pearson", label.x = -4, label.y = 4.5) +
  labs(
    title = "Correlation DESeq pseudobulk and MAST log2(PS19/WT) by cell type",
    x = "DESeq log2(PS19/WT)",
    y = "MAST log2(PS19/WT)"
  ) +
  theme_bw()

path <- paste0("../manuscript_figures/Figure_5")
saveToPDF(paste0(path, ".pdf"), width = 7.5, height = 7.5)
```
